Concluded with navigation
So far, I also added the ability for a user to sign out of the application and be redirected to the authentication page. I also have started working on the courses page. I am going to add a table of courses and a list of available lecturers in the system.